\name{extract-method}
\alias{extract-method}
\alias{[[,FlattenedArray,integer,missing-method}
\alias{[,FlattenedArray-method}
\title{Extract methods for \code{FlattenedArray} objects}
\description{
  Extract methods for objects of class
  \code{FlattenedArray}.
}
\details{
  Right now they only extract methods are implemented, no
  insertion methods. These are not particularly efficient
  either.
}

