\name{flatten-methods}
\alias{flatten}
\alias{flatten,array-method}
\alias{flatten,vector-method}
\alias{flatten-methods}
\title{Flatten array}
\arguments{
  \item{x}{Object to flatten; a vector, matrix, or array.}

  \item{default}{Value of \code{x} which are the default
  value of the array.}

  \item{drop_default}{Drop indices which have the default
  value.}
}
\value{
  Object of class \code{FlattenedArray}.
}
\description{
  Flatten matrices, arrays, and vectors.
}

