\docType{package}
\name{flatter}
\alias{flatter}
\alias{flatter-package}
\title{flatter}
\description{
  This package contains a simple class for representing
  arrays as a vector of data and indices. This is designed
  to simplify some data manipulations, and not optimized
  for high performance computing with sparse arrays.
}
\details{
  This package implements similar functionality to
  \pkg{slam}.
}

