% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.r
\name{na_type}
\alias{na_type}
\title{Given a vector, return the appropriate NA type}
\usage{
na_type(x)
}
\arguments{
\item{x}{Object}
}
\value{
A one-element vector containing the \code{NA} constant associated
with the class of \code{x}.
}
\description{
Given a vector, return the appropriate NA type
}
\seealso{
\code{\link{NA}} for a description of different missing
value constants.
}
\keyword{internal}

