% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.r, R/flatten.r
\docType{methods}
\name{flatten-methods}
\alias{flatten}
\alias{flatten,array-method}
\alias{flatten,vector-method}
\alias{flatten-methods}
\title{Flatten matrices, arrays, and vectors.}
\usage{
flatten(x, ...)

\S4method{flatten}{array}(x, default = NA, drop_default = TRUE)

\S4method{flatten}{vector}(x, ...)
}
\arguments{
\item{x}{Object to flatten; a vector, matrix, or array.}

\item{...}{passed to next method.}

\item{default}{Value of \code{x} which are the default value of the array.}

\item{drop_default}{Drop indices which have the default value.}
}
\value{
Object of class \code{FlattenedArray}.
}
\description{
Flatten matrices, arrays, and vectors.
}

